+++ packages/typespec-go/CHANGELOG.md
@@ -6,6 +6,11 @@
 
 * Updated the minimum version of `azcore` to `v1.19.0`.
 * Report a `NameCollision` diagnostic if two `const` values have the same name.
+* Updated to the latest tsp toolset.
+
+### Bugs Fixed
+
+* Fix polymorphic type result of response generation.
 
 ## 0.8.0 (2025-08-12)